# Introduce RawSource and MaskedSource instead of IndexedSource