… src/racer/codecleaner.rs
… src/racer/nameres.rs
… src/racer/scopes.rs
… src/racer/typeinf.rs