Current `IndexdSource` has some problems:
- `coords_to_point` and `point_to_coords` can be called for masked source. It's incorrect.
- We can't know it's masked or not by type

So I introduce 2 types `RawSource` and `MaskedSource`.
`RawSource` is only for coord<->point conversion and getting doc comments.
We do all operations on source codes via `MaskedSource`.
By supposing a source code is masked, we can speed up some functions.
Now I only modified `scope_start`, but some functions also need to be modified.
Here's a current benchmark result.
before(current master: 0fce5df5f567d7f36890c002de00e131651a26c1)
```
test completes_hashmap               ... bench:     722,266 ns/iter (+/- 20,725)
test completes_methods_for_file_open ... bench:     641,672 ns/iter (+/- 22,381)
test completes_methods_for_vec       ... bench: 105,745,576 ns/iter (+/- 4,441,718)
test follows_collections             ... bench:  16,464,915 ns/iter (+/- 448,679)
test follows_std                     ... bench:  21,376,878 ns/iter (+/- 631,305)
```
after: 
```
test completes_hashmap               ... bench:     696,997 ns/iter (+/- 18,631)
test completes_methods_for_file_open ... bench:     641,674 ns/iter (+/- 15,077)
test completes_methods_for_vec       ... bench:  70,758,324 ns/iter (+/- 535,415)
test follows_collections             ... bench:  14,956,695 ns/iter (+/- 179,613)
test follows_std                     ... bench:  17,238,602 ns/iter (+/- 226,426)
```